+++ internal/nix/nixos/multi.nix
@@ -11,7 +11,7 @@ let
   submodule = { config, options, name, ... }:
     let
       inherit (import ../settings.nix { inherit pkgs lib; }) format settingsModule;
-      configFile = format.generate "hercules-ci-agent${suffix}.toml" config.settings;
+      configFile = format.generate "hercules-ci-agent${suffix}.json" config.settings;
       command = "${config.package}/bin/hercules-ci-agent --config ${configFile}";
       testCommand = "${command} --test-configuration";
       suffix = if name == "" then "" else "-${name}";
@@ -33,7 +33,7 @@ let
         };
         settings = mkOption {
           description = ''
-            These settings are written to the `agent.toml` file.
+            These settings are written to the `agent.json` file.
 
             Not all settings are listed as options, can be set nonetheless.